#1154 Fix lockbox service Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13591 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/financials/src/org/opentaps/financials/domain/billing/lockbox/LockboxService.java
@@ -420,7 +420,7 @@ public class LockboxService extends Service implements LockboxServiceInterface {
                         input.put("currencyUomId", org.getPartyAcctgPreference().getBaseCurrencyUomId());
                     }
                     // receive ONLY the amount to apply, the cash discount will be accounted as an invoice adjustment
-                    input.put("amount", detail.getAmountToApply().doubleValue());
+                    input.put("amount", detail.getAmountToApply());
                     input.put("statusId", "PMNT_NOT_PAID");
                     input.put("effectiveDate", batch.getDatetimeEntered());
                     // add some references
@@ -470,7 +470,7 @@ public class LockboxService extends Service implements LockboxServiceInterface {
                             input.put("overrideGlAccountId", lockboxGlAccount.getGlAccountId());
                             // the cash discount on the LockboxItemDetail has an opposite sign:
                             // -$10 means to apply 10$ of the payment to the GL account
-                            input.put("amountApplied", detail.getCashDiscount().negate().doubleValue());
+                            input.put("amountApplied", detail.getCashDiscount().negate());
                             // call the createPaymentApplication service
                             input.put("userLogin", getUser().getOfbizUserLogin());
                             results = runSync("createPaymentApplication", input);
@@ -480,7 +480,7 @@ public class LockboxService extends Service implements LockboxServiceInterface {
                         input = new HashMap<String, Object>();
                         input.put("paymentId", paymentId);
                         input.put("invoiceId", detail.getInvoice().getInvoiceId());
-                        input.put("amountApplied", detail.getAmountToApplyToInvoice().doubleValue());
+                        input.put("amountApplied", detail.getAmountToApplyToInvoice());
                         // call the createPaymentApplication service
                         input.put("userLogin", getUser().getOfbizUserLogin());
                         results = runSync("createPaymentApplication", input);
